e2e: serial: split the node selector test
Split off the nodeselector test to their own golang source.
Since these tests don't depend on the TM scope,
de-entangle them by the `scope=container` constraint.
